Move isProducerOfSubject in utils
