+++ utils/discussions.ts
@@ -1,7 +1,7 @@
 import type { Dataservice, Dataset, Reuse } from '@datagouv/components-next'
 import { RiArticleLine, RiDatabase2Line, RiLineChartLine, RiTerminalLine } from '@remixicon/vue'
 import type { $Fetch } from 'ofetch'
-import type { DiscussionSubject, DiscussionSubjectTypes } from '~/types/discussions'
+import type { Comment, DiscussionSubject, DiscussionSubjectTypes } from '~/types/discussions'
 
 export async function getSubject(api: $Fetch, subject: DiscussionSubject): Promise<DiscussionSubjectTypes | null> {
   switch (subject.class) {
@@ -60,4 +60,16 @@ export function getDiscussionUrl(discussionId: string, subject: DiscussionSubjec
     return ''
   }
   return `${getSubjectPage(subject)}#/discussions/${discussionId}`
+}
+
+export function isProducerOfSubject(subject: DiscussionSubjectTypes, comment: Comment): boolean {
+  if (subject.owner && !comment.posted_by_organization && subject.owner.id === comment.posted_by.id) {
+    return true
+  }
+
+  if ('organization' in subject && subject.organization && comment.posted_by_organization && subject.organization.id == comment.posted_by_organization.id) {
+    return true
+  }
+
+  return false
 }
